requirements.txt
# Core RAG Framework
langchain==0.1.0
langchain-community==0.0.10

# Vector Database
qdrant-client==1.7.0

# Embeddings
sentence-transformers==2.3.1
torch==2.1.2

# LLM Provider
langchain-ollama==0.0.1

# Document Processing & Web Scraping
beautifulsoup4==4.12.3
playwright==1.41.0
lxml==5.1.0
html2text==2024.2.26

# Chunking & Text Processing
tiktoken==0.5.2
nltk==3.8.1

# Configuration
pyyaml==6.0.1
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Web Interface
streamlit==1.30.0

# Utilities
requests==2.31.0
tqdm==4.66.1

# Logging & Monitoring
loguru==0.7.2

# Development Tools
pytest==7.4.4
pytest-asyncio==0.23.3
black==24.1.1
ruff==0.1.14
mypy==1.8.0

# Type Stubs
types-requests==2.31.0.20240125
types-pyyaml==6.0.12.12
